Optimize annex_plot_generator (#205)
* Move helper functions to utils
* Rename "annual" to "yearly"
* Move settings to species-specific dicts
* Delete repeated code:
- replace the 3 for-loops by a single one over all species

* Apply formatter
* Degrade warning to info
* Add config PARISNID2025
